+++ RELEASE_NOTES.md
@@ -7,7 +7,7 @@ Release Notes
 
 - Updated all packages and framework library dependencies to .NET Core 2.2.
 - Added a new `GET_HEAD` http handler (see: [#314](https://github.com/giraffe-fsharp/Giraffe/issues/314) for more info).
-- Added two new convenience functions, `handleContext` and `handleRequest`, which can be used for creating new `HttpHandler` functions.
+- Added new convenience functions: `handleContext`, `handleContextAsync`, `handleRequest` and `handleRequestAsync`, which can be used for creating new `HttpHandler` functions.
 
 #### Bug fixes